♻️ refactor(Core): 取消迁移动画
